# tests/contract/api/openapi_snapshot.json
{
 "components": {
  "schemas": {
...
   "AdvanceRequest": {
    "additionalProperties": false,
    "properties": {
     "target_state": {
      "title": "Target State",
      "type": "string"
     }
    },
    "required": [
     "target_state"
    ],
    "title": "AdvanceRequest",
    "type": "object"
   },
...
   "CreateAgentReleaseRequest": {
    "additionalProperties": false,
    "description": "POST /agents/{id}/releases \u8bf7\u6c42\u4f53\uff1a\u4f9d\u8d56 digest \u7531\u8c03\u7528\u65b9\u63d0\u4f9b\uff1bagent_id/\nagent_version \u4ece agent \u8bb0\u5f55\u6d3e\u751f\uff08\u4e0d\u7ecf\u8bf7\u6c42\u4f53\u58f0\u660e\uff0c\u675c\u7edd\u8de8 agent \u5f15\u7528\uff09\u3002",
    "properties": {
     "approver": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Approver"
     },
     "capability_digest": {
      "title": "Capability Digest",
      "type": "string"
     },
     "eval_digests": {
      "default": [],
      "items": {
       "type": "string"
      },
      "title": "Eval Digests",
      "type": "array"
     },
     "knowledge_digest": {
      "title": "Knowledge Digest",
      "type": "string"
     },
     "memory_digest": {
      "title": "Memory Digest",
      "type": "string"
     },
     "model_digest": {
      "title": "Model Digest",
      "type": "string"
     },
     "pack_digest": {
      "title": "Pack Digest",
      "type": "string"
     },
     "policy_digest": {
      "title": "Policy Digest",
      "type": "string"
     },
     "rollback": {
      "$ref": "#/components/schemas/RollbackPolicy"
     },
     "rollout": {
      "$ref": "#/components/schemas/RolloutPolicy"
     }
    },
    "required": [
     "pack_digest",
     "model_digest",
     "knowledge_digest",
     "memory_digest",
     "capability_digest",
     "policy_digest",
     "rollout",
     "rollback"
    ],
    "title": "CreateAgentReleaseRequest",
    "type": "object"
   },
...
   "CreateReleaseRequest": {
    "additionalProperties": false,
    "properties": {
     "agent_id": {
      "format": "uuid",
      "title": "Agent Id",
      "type": "string"
     },
     "agent_version": {
      "title": "Agent Version",
      "type": "integer"
     },
     "approver": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Approver"
     },
     "capability_digest": {
      "title": "Capability Digest",
      "type": "string"
     },
     "eval_digests": {
      "default": [],
      "items": {
       "type": "string"
      },
      "title": "Eval Digests",
      "type": "array"
     },
     "knowledge_digest": {
      "title": "Knowledge Digest",
      "type": "string"
     },
     "memory_digest": {
      "title": "Memory Digest",
      "type": "string"
     },
     "model_digest": {
      "title": "Model Digest",
      "type": "string"
     },
     "pack_digest": {
      "title": "Pack Digest",
      "type": "string"
     },
     "policy_digest": {
      "title": "Policy Digest",
      "type": "string"
     },
     "rollback": {
      "$ref": "#/components/schemas/RollbackPolicy"
     },
     "rollout": {
      "$ref": "#/components/schemas/RolloutPolicy"
     }
    },
    "required": [
     "agent_id",
     "agent_version",
     "pack_digest",
     "model_digest",
     "knowledge_digest",
     "memory_digest",
     "capability_digest",
     "policy_digest",
     "rollout",
     "rollback"
    ],
    "title": "CreateReleaseRequest",
    "type": "object"
   },
...
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
...
   "ReleaseView": {
    "additionalProperties": false,
    "properties": {
     "agent_id": {
      "title": "Agent Id",
      "type": "string"
     },
     "agent_version": {
      "title": "Agent Version",
      "type": "integer"
     },
     "default_version": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Default Version"
     },
     "manifest_digest": {
      "title": "Manifest Digest",
      "type": "string"
     },
     "release_id": {
      "title": "Release Id",
      "type": "string"
     },
     "state": {
      "title": "State",
      "type": "string"
     }
    },
    "required": [
     "release_id",
     "agent_id",
     "agent_version",
     "state",
     "manifest_digest",
     "default_version"
    ],
    "title": "ReleaseView",
    "type": "object"
   },
...
   "RollbackRequest": {
    "additionalProperties": false,
    "properties": {
     "in_flight_run_ids": {
      "default": [],
      "items": {
       "format": "uuid",
       "type": "string"
      },
      "title": "In Flight Run Ids",
      "type": "array"
     },
     "to_version": {
      "title": "To Version",
      "type": "integer"
     }
    },
    "required": [
     "to_version"
    ],
    "title": "RollbackRequest",
    "type": "object"
   },
...
   "RouteRequest": {
    "additionalProperties": false,
    "properties": {
     "suspended": {
      "default": false,
      "title": "Suspended",
      "type": "boolean"
     },
     "user_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Id"
     },
     "workspace_id": {
      "anyOf": [
       {
        "format": "uuid",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Workspace Id"
     }
    },
    "title": "RouteRequest",
    "type": "object"
   },
...
  }
 },
 "info": {
  "title": "zhiwei",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/agents/{agent_id}/releases": {
   "post": {
    "operationId": "create_agent_release_api_v1_agents__agent_id__releases_post",
    "parameters": [
     {
      "in": "path",
      "name": "agent_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Agent Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateAgentReleaseRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReleaseView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Agent Release",
    "tags": [
     "agents"
    ]
   }
  },
...
  "/api/v1/releases": {
   "get": {
    "operationId": "list_releases_api_v1_releases_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/ReleaseView"
         },
         "title": "Response List Releases Api V1 Releases Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "summary": "List Releases",
    "tags": [
     "releases"
    ]
   },
   "post": {
    "operationId": "create_release_api_v1_releases_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/CreateReleaseRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReleaseView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Create Release",
    "tags": [
     "releases"
    ]
   }
  },
  "/api/v1/releases/{release_id}": {
   "get": {
    "operationId": "get_release_api_v1_releases__release_id__get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Release Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReleaseView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Release",
    "tags": [
     "releases"
    ]
   }
  },
  "/api/v1/releases/{release_id}/advance": {
   "post": {
    "operationId": "advance_release_api_v1_releases__release_id__advance_post",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Release Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/AdvanceRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReleaseView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Advance Release",
    "tags": [
     "releases"
    ]
   }
  },
  "/api/v1/releases/{release_id}/manifest": {
   "get": {
    "operationId": "get_release_manifest_api_v1_releases__release_id__manifest_get",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Release Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/ReleaseManifestView"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Get Release Manifest",
    "tags": [
     "releases"
    ]
   }
  },
  "/api/v1/releases/{release_id}/rollback": {
   "post": {
    "operationId": "rollback_release_api_v1_releases__release_id__rollback_post",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Release Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RollbackRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Rollback Release Api V1 Releases  Release Id  Rollback Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Rollback Release",
    "tags": [
     "releases"
    ]
   }
  },
  "/api/v1/releases/{release_id}/route": {
   "post": {
    "operationId": "route_release_api_v1_releases__release_id__route_post",
    "parameters": [
     {
      "in": "path",
      "name": "release_id",
      "required": true,
      "schema": {
       "format": "uuid",
       "title": "Release Id",
       "type": "string"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/RouteRequest"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "additionalProperties": true,
         "title": "Response Route Release Api V1 Releases  Release Id  Route Post",
         "type": "object"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Route Release",
    "tags": [
     "releases"
    ]
   }
  },
...
 }
}